% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datautilities.R
\name{describe}
\alias{describe}
\title{produces summary statistics for a numeric vector}
\usage{
describe(x)
}
\arguments{
\item{x}{A numeric, complex or logical vector}
}
\description{
produces summary statistics for a numeric vector
}
\examples{
x<-rnorm(100)
describe(x)
}
