% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmformat.R
\name{glmCIs}
\alias{glmCIs}
\title{Returns coefficients,  95\% confidence intervals, and p-values}
\usage{
glmCIs(x, dps = 3)
}
\arguments{
\item{x}{glm object}

\item{dps}{integer, number of decimal places desired (default=3)}
}
\value{
dataframe
}
\description{
Returns coefficients, #' 95\% confidence intervals,
and p-values from a \code{glm} object
}
\examples{
data(mtcars)
otemp<-with(mtcars, glm(mpg~+wt+vs))
glmCIs(otemp)
}
