% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeupdate.R
\name{mergeupdate}
\alias{mergeupdate}
\title{1:1 or 1:many update merge}
\usage{
mergeupdate(df1, df2, id, id1, id2, mergetype = "1:1")
}
\arguments{
\item{df1}{dataframe to be updated}

\item{df2}{dataframe with new observations or data to update}

\item{id}{common id}

\item{id1}{id for dataframe 1 (optional if common id not specified)}

\item{id2}{id for dataframe 2 (optional if common id not specified)}

\item{mergetype}{character 1:1 (default) or 1:m}
}
\value{
dataframe
}
\description{
1:1 or 1:many update merge
}
\note{
Updates values in df1 with values from df2

If df2 columns contain NAs, df1 values are retained

1 to many merges expands original dataset(df1), creates new id and duplicates values for which there are no values in second dataset

also 1:m re-sorts dataset based on id
}
\examples{
data(iris)
iris2<-iris
iris$id<-1:dim(iris)[1]
iris2<-iris
iris2$Sepal.Length<-rnorm(150)
iris2$Petal.Width[25:75]<-rnorm(51)
iris2$color="red"
iris$height="tall"
xx<-mergeupdate(iris, iris2, id="id")
data1<-data.frame(id=rep(1:10), x=1, y=10*runif(10), z=10)
data3<-data.frame(id=rep(1:10, 3), z=rnorm(30), y=100*runif(30))
mergeupdate(data1, data3, id="id", mergetype="1:m")
}
