% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datautilities.R
\name{replacedf}
\alias{replacedf}
\title{global replace of values from a data frame}
\usage{
replacedf(dat, from, to)
}
\arguments{
\item{dat}{A data frame}

\item{from}{value to replace}

\item{to}{new value}
}
\value{
A dataframe
}
\description{
global replace of values from a data frame
}
\note{
All occurences of "from" or replaced with "to"
even partial strings or individual numbers within a numeric value
}
\examples{
data(iris)
newdat<-replacedf(iris, "setosa", "set")
newdat<-replacedf(iris, 1.2, 1.25)
df<-data.frame(a=c(-99, 1, 2, 3, -99), b=c(1, 2, 3, -99, 4))
replacedf(df, -99, NA)
}
