% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datautilities.R
\name{rowsumna}
\alias{rowsumna}
\title{row sums of an array or matrix correctly accounting for NA}
\usage{
rowsumna(x)
}
\arguments{
\item{x}{a numeric array or matrix}
}
\value{
returns NA only of all are NA, otherwise returns sum of non missing value
}
\description{
row sums of an array or matrix correctly accounting for NA
}
\note{
calls sumna. Compare to rowSum(x, na.rm=TRUE) which returns 0 if all values are NA
}
\examples{
x<-c(rep(NA, 5))
y<-c(rep(NA, 3), 1, 2)
rowSums(cbind(x, y), na.rm=TRUE)
rowsumna(cbind(x, y))
library(dplyr)
xdat<-data.frame(x=c(rep(NA, 5)), y=c(rep(NA, 3), 1, 2), z=c(NA, 1:4))
xdat \%>\%
mutate(q=rowsumna(cbind(x, y, z)))
}
