% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_model.R
\name{select_model}
\alias{select_model}
\title{Backwards stepwise model  selection using AIC, accounting for missing data}
\usage{
select_model(model, include = NULL)
}
\arguments{
\item{model}{object of class "glm"}

\item{include}{character vector of variables excluded from selection}
}
\value{
"glm" object
}
\description{
Selects best-fitting model, minimizing AIC accounting for missing
values. Accounts for missing variables in two stages. 
Excludes missing observations from variables in base model, selects model.
Excludes missing observations from variables retained in selected model.
Refits model and selects model again. Then refits model to full data set.
Calls -stepAIC- from MASS library
}
\examples{
utils::data(survey, package = "MASS")
survey2<-survey
#add some missing data
survey2$Wr.Hnd[1:5]<-NA
otemp<-glm(Sex~Wr.Hnd+NW.Hnd+Fold+Pulse+Smoke+Height, data=survey2, family=binomial)
summary(otemp)
mod1<-select_model(otemp)
summary(mod1)
mod2<-select_model(otemp, include=c("Smoke", "Pulse"))
summary(mod2)
}
