% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datautilities.R
\name{substrRight}
\alias{substrRight}
\title{extract subtrings in a character vector from right}
\usage{
substrRight(x, n)
}
\arguments{
\item{x}{A character vector}

\item{n}{integer, first element from right to be extracted}
}
\value{
A character vector
}
\description{
extract subtrings in a character vector from right
}
\examples{
substrRight("testing", 3)
}
