% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabfunctions.R
\name{tabpct2}
\alias{tabpct2}
\title{Tables with percentages}
\usage{
tabpct2(row, col = NULL, collab = "N", tex = FALSE, rowlab = NULL)
}
\arguments{
\item{row}{factor or numeric vector which will appear in rows}

\item{col}{factor or numeric vector which will appear in columns}

\item{collab}{character vector of column names}

\item{tex}{output in tex format (TRUE/FALSE)}

\item{rowlab}{character vector of row names}
}
\value{
r x c matrix
}
\description{
Produces one, two or multi-way tables with percentages. Requires gdata::interleave
See tabpct3 for additional flexibility
}
\examples{
data(iris)
tabpct2(iris$Species)
iris$hilo<-cut(iris$Sepal.Width, breaks=2)
tabpct2(iris$Species, iris$hilo)
tabpct2(iris$Species, iris$hilo, collab=c("low", "Pct.", "hi", "Pct."))
tabpct2(iris$hilo, iris$Species, rowlab=c("low", "hi", "Total"), collab=c("set", "\%", "vers", "\%", "virg", "\%"))
}
