% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabfunctions.R
\name{tabpct3}
\alias{tabpct3}
\title{Tables with percentages}
\usage{
tabpct3(
  row,
  col = NULL,
  collab = NULL,
  rowlab = NULL,
  pcts = "col",
  dp = 2,
  total = TRUE,
  pval = "exact"
)
}
\arguments{
\item{row}{factor or numeric vector which will appear in rows}

\item{col}{factor or numeric vector which will appear in columns}

\item{collab}{character vector of column names}

\item{rowlab}{character vector of row names}

\item{pcts}{character string 'both', 'row', or 'col' indicating row, column or both percents (default='col')}

\item{dp}{integer indicating number of decimal places for percents (default=2)}

\item{total}{Logical indicating whether row and column totals are shown (default= TRUE)}

\item{pval}{character string indicating 'exact', 'chisq' or 'none' (default='exact')}
}
\value{
r x c matrix
}
\description{
Produces one, two or multi-way tables with percentages. Requires gdata::interleave.
For latex, html output, object can be formatted with knitr::kable
}
\examples{
library(knitr)
yn<-sample(c("Y", "N"), 100, replace=TRUE)
sex<-sample(c("M", "F"), 100, replace=TRUE)
tabpct3(yn)
tabpct3(yn, sex)
tabpct3(yn, sex, dp=1, pcts="row", rowlab=c("No", "Yes"), collab=c("Female", "Male"))
tabpct3(yn, sex, dp=1, pcts="col", rowlab=c("No", "Yes"), collab=c("Female", "Male"))
xx<-tabpct3(yn, sex, dp=1, pcts="col", rowlab=c("No", "Yes"), collab=c("Female", "Male"))
kable(xx, format="latex", caption="My Table")
kable(xx, format="simple", caption="My Table")
ynm<-sample(c("y", "n", "m"), 312, replace=TRUE)
dnum<-sample(c(1:5), 312, replace=TRUE)
tabpct3(dnum, ynm, collab=c("Maybe", "No", "Yes"), pval="chisq")
}
